backend/requirements.txt
llama-index==0.9.48
chromadb
gradio
langchain
langchain-community
python-docx
textract
PyPDF2
mysql-connector-python
ollama
transformers
sentence-transformers
python-dotenv
fastapi
uvicorn
psutil
python-jose
requests